feat: add aws config
